Update to use frida instead of mach_inject & fishhook, now works on Apple Silicon

+++ install/uninstall.sh
@@ -5,8 +5,11 @@ if [[ $EUID -ne 0 ]]; then
    exit 1
 fi
 
+if test -f "/usr/local/forceFullDesktopBar/bootstrap.dylib"; then
+    rm /usr/local/forceFullDesktopBar/bootstrap.dylib
+fi
+
 rm /usr/local/forceFullDesktopBar/forceFullDesktopBar
-rm /usr/local/forceFullDesktopBar/bootstrap.dylib
 rm /usr/local/forceFullDesktopBar/dockInjection.dylib
 rmdir /usr/local/forceFullDesktopBar
 rm /Library/LaunchDaemons/net.briankendall.forceFullDesktopBar.plist
